Restructured SpringerJSONConverter
